server/requirements.txt
fastapi
uvicorn
google-genai
langchain-google-genai
langgraph
chromadb
pypdf
python-multipart
python-dotenv
